19/10/2023 17:08
Modifié le 19/10/2023 à 17:08

+++ src/gsb/modele/Medicament.java
@@ -7,7 +7,7 @@ public class Medicament {
 	protected String composition;
 	protected String effets;
 	protected String contreIndication;
-	protected float prixEchantillon;
+	protected double prixEchantillon;
 	protected String codeFamille;
 	protected String libellefamille;
 	/**
@@ -21,7 +21,7 @@ public class Medicament {
 	 * @param libellefamille
 	 */
 	public Medicament(String depotLegal, String nomCommercial, String composition, String effets,
-			String contreIndication, float prixEchantillon, String codeFamille, String libellefamille) {
+			String contreIndication, double prixEchantillon, String codeFamille, String libellefamille) {
 		this.depotLegal = depotLegal;
 		this.nomCommercial = nomCommercial;
 		this.composition = composition;
@@ -94,13 +94,13 @@ public class Medicament {
 	/**
 	 * @return the prixEchantillon
 	 */
-	public float getPrixEchantillon() {
+	public double getPrixEchantillon() {
 		return prixEchantillon;
 	}
 	/**
 	 * @param prixEchantillon the prixEchantillon to set
 	 */
-	public void setPrixEchantillon(float prixEchantillon) {
+	public void setPrixEchantillon(double prixEchantillon) {
 		this.prixEchantillon = prixEchantillon;
 	}
 	/**
